agape-platform-backend/requirements.txt
flask==2.2.3
flask-pymongo==2.3.0
flask-jwt-extended==4.4.4
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
bcrypt==4.0.1
dnspython==2.3.0
email-validator==1.3.1
gunicorn==20.1.0
Werkzeug==2.2.3
